+++ libs/sra/sratbl-cmn.c
@@ -1214,11 +1214,11 @@ enum sfa_path_type_id CC sfa_path_type_db ( const char *path )
 #undef MATCH
 
 static
-int CC sfa_path_cmp ( const void **_a, const void **_b, void * ignored )
+int64_t CC sfa_path_cmp ( const void **_a, const void **_b, void * ignored )
 {
     const reorder_t * a = *_a;
     const reorder_t * b = *_b;
-    int ret;
+    int64_t ret;
 
     DEBUG_SORT(("%s enter\t%s %s %lu \t%s %s %lu", __func__, 
                 a->path, sfa_path_type_id[a->type_id + 1], a->size,
